separated pages from components
