+++ src/components/index.js
@@ -1,29 +1,8 @@
-import Admin from './Admin'
-import Home from './Home'
 import NavBar from './NavBar'
-import Account from './Account'
-import SignIn from './SignIn'
-import SignUp from './SignUp'
-import Showroom from './Showroom'
-import PasswordReset from './PasswordReset'
 import LoggedInHome from './LoggedInHome'
 import LoggedOutHome from './LoggedOutHome'
 import Footer from './Footer'
 import GenericLink from './GenericLink'
 import FormInputAndLabel from './FormInputAndLabel'
 
-export {
-  Admin,
-  Home,
-  NavBar,
-  Account,
-  SignIn,
-  SignUp,
-  Showroom,
-  PasswordReset,
-  LoggedInHome,
-  LoggedOutHome,
-  Footer,
-  GenericLink,
-  FormInputAndLabel
-}
+export { NavBar, LoggedInHome, LoggedOutHome, Footer, GenericLink, FormInputAndLabel }
