+++ src/pages/index.js
@@ -1,0 +1,9 @@
+import Account from '../pages/Account'
+import Admin from '../pages/Admin'
+import Home from '../pages/Home'
+import PasswordReset from '../pages/PasswordReset'
+import Showroom from '../pages/Showroom'
+import SignIn from '../pages/SignIn'
+import SignUp from '../pages/SignUp'
+
+export { Account, Admin, Home, PasswordReset, Showroom, SignIn, SignUp }
